refactor(canvas): finish chat root cleanup
