+++ apps/canvas-workspace/src/renderer/src/components/chat/SessionTitle/index.test.tsx
@@ -2,7 +2,7 @@
 import { act } from 'react';
 import { createRoot } from 'react-dom/client';
 import { describe, expect, it } from 'vitest';
-import { SessionTitle } from './SessionTitle';
+import { SessionTitle } from '.';
 
 (globalThis as { IS_REACT_ACT_ENVIRONMENT?: boolean }).IS_REACT_ACT_ENVIRONMENT = true;
